requirements.txt
requests~=2.28.1
pyTelegramBotAPI~=4.7.1
python-dotenv~=0.21.0
